refactor: consolidate ConfigMap override test helpers and fix copyright
- Create shared test_util/configmap_override.py with common helper functions
- Update copyright year to 2026 in all test files
- Replace em dashes with regular hyphens in step comments
- Remove duplicate code from cilium, metallb, localpv, and metrics-server tests
- Delete test_cilium_devices_override.py (no longer needed)
