+++ tests/integration/tests/test_localpv_configmap_override.py
@@ -1,5 +1,5 @@
 #
-# Copyright 2025 Canonical, Ltd.
+# Copyright 2026 Canonical, Ltd.
 #
 """Integration test for the LocalPV ConfigMap override feature.
 
@@ -21,8 +21,7 @@ import logging
 from typing import List
 
 import pytest
-import yaml
-from test_util import config, harness, tags, util
+from test_util import config, harness, tags, util, configmap_override
 
 LOG = logging.getLogger(__name__)
 
@@ -35,60 +34,6 @@ HELM_NAMESPACE = "kube-system"
 DEFAULT_RECLAIM_POLICY = "Retain"
 
 
-def _apply_override_configmap(instance: harness.Instance, values_yaml: str):
-    manifest = (
-        f"apiVersion: v1\n"
-        f"kind: ConfigMap\n"
-        f"metadata:\n"
-        f"  name: {OVERRIDE_CM_NAME}\n"
-        f"  namespace: {OVERRIDE_CM_NAMESPACE}\n"
-        f"data:\n"
-        f"  values: |\n"
-    )
-    for line in values_yaml.splitlines():
-        manifest += f"    {line}\n"
-    instance.exec(
-        ["k8s", "kubectl", "apply", "-f", "-"],
-        input=manifest.encode(),
-    )
-
-
-def _delete_override_configmap(instance: harness.Instance):
-    instance.exec(
-        [
-            "k8s",
-            "kubectl",
-            "delete",
-            "configmap",
-            OVERRIDE_CM_NAME,
-            "-n",
-            OVERRIDE_CM_NAMESPACE,
-            "--ignore-not-found",
-        ],
-    )
-
-
-def _helm_values_cmd() -> List[str]:
-    return [
-        "k8s",
-        "helm",
-        "get",
-        "values",
-        HELM_RELEASE,
-        "--namespace",
-        HELM_NAMESPACE,
-        "--output",
-        "yaml",
-    ]
-
-
-def _parse_helm_stdout(p) -> dict:
-    raw = p.stdout
-    if isinstance(raw, bytes):
-        raw = raw.decode()
-    return yaml.safe_load(raw) or {}
-
-
 def _wait_for_reclaim_policy(
     instance: harness.Instance,
     expected: str,
@@ -98,7 +43,7 @@ def _wait_for_reclaim_policy(
     """Poll Helm values until storageClass.reclaimPolicy matches expected."""
 
     def _policy_matches(p) -> bool:
-        values = _parse_helm_stdout(p)
+        values = configmap_override.parse_helm_stdout(p)
         actual = values.get("storageClass", {}).get("reclaimPolicy")
         LOG.info("Helm storageClass.reclaimPolicy: %s (want %s)", actual, expected)
         return actual == expected
@@ -106,7 +51,7 @@ def _wait_for_reclaim_policy(
     util.stubbornly(retries=retries, delay_s=delay_s).on(instance).until(
         _policy_matches
     ).exec(
-        _helm_values_cmd(),
+        configmap_override.helm_values_cmd(HELM_RELEASE, HELM_NAMESPACE),
     )
 
 
@@ -119,34 +64,40 @@ def test_localpv_configmap_override(instances: List[harness.Instance]):
     try:
         util.wait_until_k8s_ready(instance, [instance])
 
-        # --- Step 1: Apply initial ConfigMap override ---
+        # -- Step 1: Apply initial ConfigMap override --
         # Override storageClass.reclaimPolicy from the k8sd default (Retain) to Delete.
         LOG.info(
             "Applying LocalPV override ConfigMap with storageClass.reclaimPolicy=Delete"
         )
-        _apply_override_configmap(
+        configmap_override.apply_override_configmap(
             instance,
+            OVERRIDE_CM_NAME,
+            OVERRIDE_CM_NAMESPACE,
             "storageClass:\n  reclaimPolicy: Delete\n",
         )
 
         LOG.info("Waiting for Helm to reflect storageClass.reclaimPolicy=Delete")
         _wait_for_reclaim_policy(instance, expected="Delete")
 
-        # --- Step 2: Update the ConfigMap override ---
+        # -- Step 2: Update the ConfigMap override --
         LOG.info(
             "Updating LocalPV override ConfigMap with storageClass.reclaimPolicy=Recycle"
         )
-        _apply_override_configmap(
+        configmap_override.apply_override_configmap(
             instance,
+            OVERRIDE_CM_NAME,
+            OVERRIDE_CM_NAMESPACE,
             "storageClass:\n  reclaimPolicy: Recycle\n",
         )
 
         LOG.info("Waiting for Helm to reflect storageClass.reclaimPolicy=Recycle")
         _wait_for_reclaim_policy(instance, expected="Recycle")
 
-        # --- Step 3: Delete the ConfigMap and verify revert to k8sd defaults ---
+        # -- Step 3: Delete the ConfigMap and verify revert to k8sd defaults --
         LOG.info("Deleting LocalPV override ConfigMap")
-        _delete_override_configmap(instance)
+        configmap_override.delete_override_configmap(
+            instance, OVERRIDE_CM_NAME, OVERRIDE_CM_NAMESPACE
+        )
 
         LOG.info(
             "Waiting for storageClass.reclaimPolicy to revert to k8sd default (%s)",
@@ -155,4 +106,6 @@ def test_localpv_configmap_override(instances: List[harness.Instance]):
         _wait_for_reclaim_policy(instance, expected=DEFAULT_RECLAIM_POLICY)
 
     finally:
-        _delete_override_configmap(instance)
+        configmap_override.delete_override_configmap(
+            instance, OVERRIDE_CM_NAME, OVERRIDE_CM_NAMESPACE
+        )
