+++ tests/integration/tests/test_metrics_server_configmap_override.py
@@ -1,5 +1,5 @@
 #
-# Copyright 2025 Canonical, Ltd.
+# Copyright 2026 Canonical, Ltd.
 #
 """Integration test for the metrics-server ConfigMap override feature.
 
@@ -19,8 +19,7 @@ import logging
 from typing import List
 
 import pytest
-import yaml
-from test_util import config, harness, tags, util
+from test_util import config, harness, tags, util, configmap_override
 
 LOG = logging.getLogger(__name__)
 
@@ -28,60 +27,6 @@ OVERRIDE_CM_NAME = "k8sd-metrics-server-values"
 OVERRIDE_CM_NAMESPACE = "kube-system"
 HELM_RELEASE = "metrics-server"
 HELM_NAMESPACE = "kube-system"
-
-
-def _apply_override_configmap(instance: harness.Instance, values_yaml: str):
-    manifest = (
-        f"apiVersion: v1\n"
-        f"kind: ConfigMap\n"
-        f"metadata:\n"
-        f"  name: {OVERRIDE_CM_NAME}\n"
-        f"  namespace: {OVERRIDE_CM_NAMESPACE}\n"
-        f"data:\n"
-        f"  values: |\n"
-    )
-    for line in values_yaml.splitlines():
-        manifest += f"    {line}\n"
-    instance.exec(
-        ["k8s", "kubectl", "apply", "-f", "-"],
-        input=manifest.encode(),
-    )
-
-
-def _delete_override_configmap(instance: harness.Instance):
-    instance.exec(
-        [
-            "k8s",
-            "kubectl",
-            "delete",
-            "configmap",
-            OVERRIDE_CM_NAME,
-            "-n",
-            OVERRIDE_CM_NAMESPACE,
-            "--ignore-not-found",
-        ],
-    )
-
-
-def _helm_values_cmd() -> List[str]:
-    return [
-        "k8s",
-        "helm",
-        "get",
-        "values",
-        HELM_RELEASE,
-        "--namespace",
-        HELM_NAMESPACE,
-        "--output",
-        "yaml",
-    ]
-
-
-def _parse_helm_stdout(p) -> dict:
-    raw = p.stdout
-    if isinstance(raw, bytes):
-        raw = raw.decode()
-    return yaml.safe_load(raw) or {}
 
 
 def _wait_for_replicas(
@@ -93,7 +38,7 @@ def _wait_for_replicas(
     """Poll Helm values until replicas matches expected."""
 
     def _replicas_match(p) -> bool:
-        values = _parse_helm_stdout(p)
+        values = configmap_override.parse_helm_stdout(p)
         actual = values.get("replicas")
         LOG.info("Helm replicas: %s (want %s)", actual, expected)
         return actual == expected
@@ -101,7 +46,7 @@ def _wait_for_replicas(
     util.stubbornly(retries=retries, delay_s=delay_s).on(instance).until(
         _replicas_match
     ).exec(
-        _helm_values_cmd(),
+        configmap_override.helm_values_cmd(HELM_RELEASE, HELM_NAMESPACE),
     )
 
 
@@ -113,7 +58,7 @@ def _wait_for_replicas_absent(
     """Poll Helm values until replicas is absent (override reverted)."""
 
     def _replicas_absent(p) -> bool:
-        values = _parse_helm_stdout(p)
+        values = configmap_override.parse_helm_stdout(p)
         absent = "replicas" not in values
         LOG.info("Helm values 'replicas' absent: %s (want True)", absent)
         return absent
@@ -121,7 +66,7 @@ def _wait_for_replicas_absent(
     util.stubbornly(retries=retries, delay_s=delay_s).on(instance).until(
         _replicas_absent
     ).exec(
-        _helm_values_cmd(),
+        configmap_override.helm_values_cmd(HELM_RELEASE, HELM_NAMESPACE),
     )
 
 
@@ -134,28 +79,42 @@ def test_metrics_server_configmap_override(instances: List[harness.Instance]):
     try:
         util.wait_until_k8s_ready(instance, [instance])
 
-        # --- Step 1: Apply initial ConfigMap override ---
+        # -- Step 1: Apply initial ConfigMap override --
         # Override replicas (k8sd does not set this, so it will appear in
         # helm get values only when explicitly overridden and disappear on delete).
         LOG.info("Applying metrics-server override ConfigMap with replicas=2")
-        _apply_override_configmap(instance, "replicas: 2\n")
+        configmap_override.apply_override_configmap(
+            instance,
+            OVERRIDE_CM_NAME,
+            OVERRIDE_CM_NAMESPACE,
+            "replicas: 2\n",
+        )
 
         LOG.info("Waiting for Helm to reflect replicas=2")
         _wait_for_replicas(instance, expected=2)
 
-        # --- Step 2: Update the ConfigMap override ---
+        # -- Step 2: Update the ConfigMap override --
         LOG.info("Updating metrics-server override ConfigMap with replicas=3")
-        _apply_override_configmap(instance, "replicas: 3\n")
+        configmap_override.apply_override_configmap(
+            instance,
+            OVERRIDE_CM_NAME,
+            OVERRIDE_CM_NAMESPACE,
+            "replicas: 3\n",
+        )
 
         LOG.info("Waiting for Helm to reflect replicas=3")
         _wait_for_replicas(instance, expected=3)
 
-        # --- Step 3: Delete the ConfigMap and verify revert ---
+        # -- Step 3: Delete the ConfigMap and verify revert --
         LOG.info("Deleting metrics-server override ConfigMap")
-        _delete_override_configmap(instance)
+        configmap_override.delete_override_configmap(
+            instance, OVERRIDE_CM_NAME, OVERRIDE_CM_NAMESPACE
+        )
 
         LOG.info("Waiting for 'replicas' to be absent from Helm values")
         _wait_for_replicas_absent(instance)
 
     finally:
-        _delete_override_configmap(instance)
+        configmap_override.delete_override_configmap(
+            instance, OVERRIDE_CM_NAME, OVERRIDE_CM_NAMESPACE
+        )
