+++ tests/integration/tox.ini
@@ -47,6 +47,7 @@ commands =
         --disable-warnings \
         {posargs:{toxinidir}/tests}
 passenv =
+    PATH
     TEST_*
 
 [testenv:pytest-benchmark]
